requirement.txt
numpy
scipy
pandas
Flask
scikit-learn
scikit-surprise
keras
tensorflow